billing/requirements.txt
psycopg2-binary==2.9.3
python-dotenv==0.20.0
gunicorn==20.1.0
django-split-settings==1.1.0
celery==5.2.7
django-celery-beat==2.3.0
django-celery-results==2.2.0
celery_once==3.0.1
djangorestframework==3.13.1
django-filter==22.1
drf-yasg==1.21.3
pydantic==1.10.0
Django==4
django-jazzmin==2.5.0
django-extensions==3.2.1
ipython==8.5.0
yookassa-payout==2.4.0
yookassa==2.3.2
factory-boy==3.2.1
yookassa
kafka-python==2.0.2
